Implement NextAuth.js email login
