--- app/(training-app)/register/page.tsx
@@ -1,5 +1,0 @@
-import AuthForm from '@/components/AuthForm'
-
-export default async function Register() {
-  return <AuthForm mode="register" />
-}
