+++ app/api/user/[slug]/route.ts
@@ -14,8 +14,7 @@ export async function GET(
       createdAt: true,
       updatedAt: true,
       email: true,
-      firstName: true,
-      lastName: true,
+      name: true,
       workouts: {
         where: {
           deleted: false,
