+++ app/profile/page.tsx
@@ -10,7 +10,7 @@ export default async function Profile() {
     <Container>
       <section>
         <h1 className="text-6xl font-bold leading-tight tracking-tighter md:pr-8 md:text-8xl">
-          Hello{user && `, ${user.firstName}`}.
+          Hello{user && `, ${user.name}`}.
         </h1>
       </section>
       {user && (
