+++ app/workout/[slug]/page.tsx
