+++ components/calendar/Calendar.tsx
@@ -1,21 +1,27 @@
 'use client'
 
-import React from 'react'
+import React, {useEffect} from 'react'
 import {CalendarMobile} from '@/components/calendar/CalendarMobile'
 import {CalendarMedium} from '@/components/calendar/CalendarMedium'
 import {CalendarHeading} from '@/components/calendar/CalendarHeading'
 import {CalendarGrid} from '@/components/calendar/CalendarGrid'
 import {useCalendarData} from '@/hooks'
-import Providers from '@/app/(training-app)/Providers'
+import Providers from '@/app/Providers'
 import {SerialisedUser} from '@/lib/users'
 import ClientWrapper from '@/components/calendar/ClientWrapper'
 import {Sidebar} from '@/components/calendar/Sidebar'
+import {useSession} from 'next-auth/react'
 
 const SidebarMemo = React.memo(Sidebar)
 
 export function Calendar({user}: {user: SerialisedUser}) {
   const {monthData, year, month, setYear, setMonth} = useCalendarData()
   const isAdmin = user.admin
+  const {data: session, status} = useSession()
+
+  useEffect(() => {
+    console.log({session, status})
+  }, [session, status])
 
   return (
     <Providers>
