+++ hooks/useGetClientsTableData.ts
@@ -5,7 +5,7 @@ import {getUsersWithWorkouts} from '@/lib/api'
 import {SerialisedUser, SerialisedUserKey, sortUsers} from '@/lib/users'
 
 export function useGetClientsTableData() {
-  const [sortCol, setSortCol] = useState<SerialisedUserKey>('firstName')
+  const [sortCol, setSortCol] = useState<SerialisedUserKey>('name')
   const [users, setUsers] = useState<SerialisedUser[] | undefined>()
 
   const fetchUsers = useCallback(async () => {
