+++ package.json
@@ -28,7 +28,8 @@
     "@contentful/rich-text-types": "^16.0.2",
     "@headlessui/react": "^1.7.13",
     "@heroicons/react": "^2.0.16",
-    "@prisma/client": "^4.11.0",
+    "@next-auth/prisma-adapter": "^1.0.6",
+    "@prisma/client": "^4.13.0",
     "@types/node": "18.14.4",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
@@ -40,8 +41,9 @@
     "eslint-config-next": "^13.3.1",
     "jose": "^4.13.1",
     "next": "^13.3.1",
+    "next-auth": "^4.22.1",
     "nodemailer": "^6.9.1",
-    "prisma": "^4.11.0",
+    "prisma": "^4.13.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "server-only": "^0.0.1",
