+++ prisma/schema.prisma
@@ -11,16 +11,20 @@ datasource db {
 }
 
 model User {
-  id        String   @id @default(uuid())
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-
-  email     String    @unique
-  password  String
   admin     Boolean   @default(false)
-  firstName String?
-  lastName  String?
+  createdAt DateTime  @default(now())
+  password  String
+  updatedAt DateTime  @updatedAt
   workouts  Workout[]
+
+  // Auth.js fields
+  accounts      Account[]
+  email         String    @unique
+  emailVerified DateTime?
+  id            String    @id @default(uuid())
+  image         String?
+  name          String?
+  sessions      Session[]
 }
 
 model Workout {
@@ -50,4 +54,39 @@ enum WORKOUT_STATUS {
 enum WORKOUT_TYPE {
   TRAINING
   APPOINTMENT
+}
+
+model Account {
+  id                String  @id @default(cuid())
+  userId            String
+  type              String
+  provider          String
+  providerAccountId String
+  refresh_token     String? @db.Text
+  access_token      String? @db.Text
+  expires_at        Int?
+  token_type        String?
+  scope             String?
+  id_token          String? @db.Text
+  session_state     String?
+
+  user User @relation(fields: [userId], references: [id], onDelete: Cascade)
+
+  @@unique([provider, providerAccountId])
+}
+
+model Session {
+  id           String   @id @default(cuid())
+  sessionToken String   @unique
+  userId       String
+  expires      DateTime
+  user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+}
+
+model VerificationToken {
+  identifier String
+  token      String   @unique
+  expires    DateTime
+
+  @@unique([identifier, token])
 }
